LayoutTests/media/media-audio-track-expected.txt


RUN(mediaElement.src = "content/short--1-track.webm")

EVENT(loadedmetadata)
EXPECTED (video.audioTracks.length == '1') OK
EXPECTED (video.audioTracks[0].enabled == 'true') OK
EXPECTED (video.audioTracks[0].id == '18446744073709551615') OK
END OF TEST
